second commit

+++ WebProject/Models/Movies.cs
@@ -16,7 +16,7 @@ namespace WebProject.Models
         [Required]
         [Display(Name="Number in Stock")]
         [Range(1, 20)]
-        public int NumberInStock { get; set; }
+        public byte NumberInStock { get; set; }
       
         
         public Genre Genre { get; set; }
